Initial MVC, and request, reponse boundary.
Initial classes created for model, view, controller. A request and a response object also created for clean encapsulation of requests and reponses. A scenario created: when is search is made for a given location, the presenter should present the response. A unit test is added for this scenario. (Unit testing must be enabled from 'Build Variants' and the test run as a gradle test).