Convert a Venue from JSON to a POJO.
Take a venue object from an API venues search on the Foursquare Developers page, and add this JSON in our test resources. Implement a Reader for reading the JSON, and converters for JSON and for a Venue item. Write a Unit test to check that the given JSON is converted into Venue object containing the expected data. The test/res cannot be read directly (due to a limitation in the Android Studio testing framework), so an additional workaround build.gradle file had to be added.