Finish CardView layout, tighten up presentation and animations.
Layout a bit more polished: can see default text properly, semi-transparent grey background added so that text can still be read on lighter backgrounds (not perfect solution, but good enough).
Retrfofit throw exception when it gets an error code (e.g. 400 Bad Request), need to handle this gracefully and pass it back to the user in an intelligible way.
Adapter and Holder implemented for Cardviews with Picasso added for image loading.